Restructured gamePlay object
such that it has a second class - which organizes a turn a little bit more modular.

Key differences:
It now updates the gameState and documents gamePlay in independent functions. The point of this is that the updateGameState function can be passed to the agent so they can sample post-state outcomes when they are choosing which play to make.