lineValue agents!
I created two new agents, one is the "best line agent" which simply measures what sequential plays can be made on it's own line, and plays the dominoe with the highest value (or a double, or an off line play that has more value than it's own line).

I also created a lineValueAgent that uses a CNN + MLP network to measure the value at each state, which includes hand-crafted information about the line possibilities.