Refactored value agents
Now there's a top level valueAgent class of which the special valueAgent types are children. DRY!!!! There might be a few logical issues to deal with (for example, handling how the networks vary about output dimensions -- right now the basic valueNetwork estimates the score of every player, where the lineValueNetwork only estimates the score of the agent).