Rule change -- numDominoes per turn
Before, the code was set up to cycle the extra dominoes by the number of players with extras. But this is problematic, because if the number of players with extra dominoes is a factor of the total number of players, then the same groups of players would get the same number each turn. It's better to use all groups of players - so now it just increments by 1 each turn.