Added Persistent Agent
Along with a slight rework of the select play code for the best line agent (which is easier to read and not slower). It also helps the persistent agent overwrite just the key methods and not rewrite lots of code.

I also added an experiment that compares the performance of basicLineAgents and persistentLineAgents.